Move TransactionRequest and TransactionResponse to separate package.
