+++ transaction-service-api/src/main/java/pl/adamboguszewski/transactionservice/service/api/Transaction/TransactionResponse.java
@@ -1,4 +1,4 @@
-package pl.adamboguszewski.transactionservice.service.api;
+package pl.adamboguszewski.transactionservice.service.api.Transaction;
 
 // [TODO]: This is application response, implement info that should be returned to application user after you create and add transaction to database (we don't have db yet but nvm)
 // [TODO]: Also look for a more proper class name to fit the request
