+++ transaction-service/src/main/java/pl/adamboguszewski/transactionservice/application/ApplicationController.java
@@ -3,7 +3,7 @@ package pl.adamboguszewski.transactionservice.application;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import pl.adamboguszewski.transactionservice.service.api.TransactionRequest;
+import pl.adamboguszewski.transactionservice.service.api.Transaction.TransactionRequest;
 
 @RestController
 @RequestMapping
